grsec: Apply grsecurity-2.9.1-2.6.32.59-201206182053.patch
commit b01fa4c83844011288da7b8a128d78548b078519
Merge: 0521c1e 4d3e3fb
Author: Brad Spengler <spender@grsecurity.net>
Date:   Mon Jun 18 20:22:09 2012 -0400

    Merge branch 'pax-stable' into grsec-stable

commit 4d3e3fb407dbc2e25951184f26a6d74658423414
Author: Brad Spengler <spender@grsecurity.net>
Date:   Mon Jun 18 20:21:50 2012 -0400

    Update to pax-linux-2.6.32.59-test175.patch

 arch/x86/include/asm/uaccess_32.h |    8 ++++----
 arch/x86/include/asm/uaccess_64.h |    8 ++++----
 tools/gcc/size_overflow_plugin.c  |    2 +-
 3 files changed, 9 insertions(+), 9 deletions(-)

Signature-tree: cb2b8e1a22c59985be829cd58e694cef6baf6d64

+++ arch/x86/include/asm/uaccess_32.h
@@ -243,9 +243,9 @@ extern void copy_from_user_overflow(void)
 static __always_inline unsigned long __must_check
 copy_to_user(void __user *to, const void *from, unsigned long n)
 {
-	int sz = __compiletime_object_size(from);
+	size_t sz = __compiletime_object_size(from);
 
-	if (unlikely(sz != -1 && sz < n))
+	if (unlikely(sz != (size_t)-1 && sz < n))
 		copy_to_user_overflow();
 	else if (access_ok(VERIFY_WRITE, to, n))
 		n = __copy_to_user(to, from, n);
@@ -271,9 +271,9 @@ copy_to_user(void __user *to, const void *from, unsigned long n)
 static __always_inline unsigned long __must_check
 copy_from_user(void *to, const void __user *from, unsigned long n)
 {
-	int sz = __compiletime_object_size(to);
+	size_t sz = __compiletime_object_size(to);
 
-	if (unlikely(sz != -1 && sz < n))
+	if (unlikely(sz != (size_t)-1 && sz < n))
 		copy_from_user_overflow();
 	else if (access_ok(VERIFY_READ, from, n))
 		n = __copy_from_user(to, from, n);
