+++ arch/x86/include/asm/uaccess_64.h
@@ -43,7 +43,7 @@ extern void copy_from_user_overflow(void)
 static __always_inline __must_check
 unsigned long __copy_from_user(void *dst, const void __user *src, unsigned long size)
 {
-	int sz = __compiletime_object_size(dst);
+	size_t sz = __compiletime_object_size(dst);
 	unsigned ret = 0;
 
 	might_fault();
@@ -56,7 +56,7 @@ unsigned long __copy_from_user(void *dst, const void __user *src, unsigned long 
 		return size;
 #endif
 
-	if (unlikely(sz != -1 && sz < size)) {
+	if (unlikely(sz != (size_t)-1 && sz < size)) {
 		copy_from_user_overflow();
 		return size;
 	}
@@ -116,7 +116,7 @@ unsigned long __copy_from_user(void *dst, const void __user *src, unsigned long 
 static __always_inline __must_check
 unsigned long __copy_to_user(void __user *dst, const void *src, unsigned long size)
 {
-	int sz = __compiletime_object_size(dst);
+	size_t sz = __compiletime_object_size(dst);
 	unsigned ret = 0;
 
 	might_fault();
@@ -131,7 +131,7 @@ unsigned long __copy_to_user(void __user *dst, const void *src, unsigned long si
 		return size;
 #endif
 
-	if (unlikely(sz != -1 && sz < size)) {
+	if (unlikely(sz != (size_t)-1 && sz < size)) {
 		copy_to_user_overflow();
 		return size;
 	}
